+++ src/RcppExports.cpp
@@ -53,16 +53,28 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// ud_convert
-NumericVector ud_convert(NumericVector val, std::string from, std::string to);
-RcppExport SEXP _units_ud_convert(SEXP valSEXP, SEXP fromSEXP, SEXP toSEXP) {
+// ud_convertible
+LogicalVector ud_convertible(std::string from, std::string to);
+RcppExport SEXP _units_ud_convertible(SEXP fromSEXP, SEXP toSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< std::string >::type from(fromSEXP);
+    Rcpp::traits::input_parameter< std::string >::type to(toSEXP);
+    rcpp_result_gen = Rcpp::wrap(ud_convertible(from, to));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ud_convert_doubles
+NumericVector ud_convert_doubles(NumericVector val, std::string from, std::string to);
+RcppExport SEXP _units_ud_convert_doubles(SEXP valSEXP, SEXP fromSEXP, SEXP toSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type val(valSEXP);
     Rcpp::traits::input_parameter< std::string >::type from(fromSEXP);
     Rcpp::traits::input_parameter< std::string >::type to(toSEXP);
-    rcpp_result_gen = Rcpp::wrap(ud_convert(val, from, to));
+    rcpp_result_gen = Rcpp::wrap(ud_convert_doubles(val, from, to));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -157,18 +169,6 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< SEXP >::type unit(unitSEXP);
     rcpp_result_gen = Rcpp::wrap(R_ut_get_symbol(unit));
-    return rcpp_result_gen;
-END_RCPP
-}
-// R_ut_are_convertible
-LogicalVector R_ut_are_convertible(SEXP a, SEXP b);
-RcppExport SEXP _units_R_ut_are_convertible(SEXP aSEXP, SEXP bSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< SEXP >::type a(aSEXP);
-    Rcpp::traits::input_parameter< SEXP >::type b(bSEXP);
-    rcpp_result_gen = Rcpp::wrap(R_ut_are_convertible(a, b));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -298,7 +298,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_units_ud_init", (DL_FUNC) &_units_ud_init, 1},
     {"_units_ud_set_encoding", (DL_FUNC) &_units_ud_set_encoding, 1},
     {"_units_ud_compare", (DL_FUNC) &_units_ud_compare, 4},
-    {"_units_ud_convert", (DL_FUNC) &_units_ud_convert, 3},
+    {"_units_ud_convertible", (DL_FUNC) &_units_ud_convertible, 2},
+    {"_units_ud_convert_doubles", (DL_FUNC) &_units_ud_convert_doubles, 3},
     {"_units_ud_map_names", (DL_FUNC) &_units_ud_map_names, 2},
     {"_units_ud_unmap_names", (DL_FUNC) &_units_ud_unmap_names, 1},
     {"_units_ud_map_symbols", (DL_FUNC) &_units_ud_map_symbols, 2},
@@ -308,7 +309,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_units_R_ut_new_dimensionless_unit", (DL_FUNC) &_units_R_ut_new_dimensionless_unit, 0},
     {"_units_R_ut_get_name", (DL_FUNC) &_units_R_ut_get_name, 1},
     {"_units_R_ut_get_symbol", (DL_FUNC) &_units_R_ut_get_symbol, 1},
-    {"_units_R_ut_are_convertible", (DL_FUNC) &_units_R_ut_are_convertible, 2},
     {"_units_R_ut_scale", (DL_FUNC) &_units_R_ut_scale, 2},
     {"_units_R_ut_offset", (DL_FUNC) &_units_R_ut_offset, 2},
     {"_units_R_ut_multiply", (DL_FUNC) &_units_R_ut_multiply, 2},